# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A REST API for tracking workouts and managing users.",
        "title": "Workout Tracker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/workouts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of all workouts in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workouts"
                ],
                "summary": "Get all workouts",
                "responses": {
                    "200": {
                        "description": "List of workouts",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.WorkoutResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new workout with exercises for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workouts"
                ],
                "summary": "Create a new workout",
                "parameters": [
                    {
                        "description": "Workout data",
                        "name": "workout",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/store.Workout"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Workout created successfully",
                        "schema": {
                            "$ref": "#/definitions/api.WorkoutResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/workouts/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a specific workout and its exercises by workout ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workouts"
                ],
                "summary": "Get workout by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workout ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Workout details",
                        "schema": {
                            "$ref": "#/definitions/api.WorkoutResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid workout ID",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Workout not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing workout and its exercises (only by the owner)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workouts"
                ],
                "summary": "Update workout",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workout ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated workout data",
                        "name": "workout",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/store.Workout"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Workout updated successfully",
                        "schema": {
                            "$ref": "#/definitions/api.WorkoutResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - not the owner",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Workout not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a workout by ID (only by the owner)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workouts"
                ],
                "summary": "Delete workout",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workout ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Workout deleted successfully"
                    },
                    "400": {
                        "description": "Invalid workout ID",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - not the owner",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Workout not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message",
                    "type": "string",
                    "example": "Invalid request payload"
                }
            }
        },
...
        "api.WorkoutEntryResponse": {
            "type": "object",
            "properties": {
                "duration_seconds": {
                    "description": "Duration in seconds",
                    "type": "integer",
                    "example": 60
                },
                "exercise_name": {
                    "description": "Name of the exercise",
                    "type": "string",
                    "example": "Push ups"
                },
                "id": {
                    "description": "Entry ID",
                    "type": "integer",
                    "example": 1
                },
                "notes": {
                    "description": "Additional notes",
                    "type": "string",
                    "example": "Good form maintained"
                },
                "order_index": {
                    "description": "Order of exercise in workout",
                    "type": "integer",
                    "example": 1
                },
                "reps": {
                    "description": "Number of repetitions",
                    "type": "integer",
                    "example": 15
                },
                "sets": {
                    "description": "Number of sets",
                    "type": "integer",
                    "example": 3
                },
                "weight": {
                    "description": "Weight in kg",
                    "type": "number",
                    "example": 75.5
                }
            }
        },
        "api.WorkoutResponse": {
            "type": "object",
            "properties": {
                "calories_burned": {
                    "description": "Calories burned",
                    "type": "integer",
                    "example": 350
                },
                "created_at": {
                    "description": "Creation timestamp",
                    "type": "string",
                    "example": "2024-01-01T12:00:00Z"
                },
                "description": {
                    "description": "Workout description",
                    "type": "string",
                    "example": "High intensity cardio"
                },
                "duration_minutes": {
                    "description": "Duration in minutes",
                    "type": "integer",
                    "example": 45
                },
                "entries": {
                    "description": "List of workout exercises",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.WorkoutEntryResponse"
                    }
                },
                "id": {
                    "description": "Workout ID",
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "description": "Workout title",
                    "type": "string",
                    "example": "Morning Cardio"
                },
                "updated_at": {
                    "description": "Last update timestamp",
                    "type": "string",
                    "example": "2024-01-01T12:00:00Z"
                },
                "user_id": {
                    "description": "User ID who owns the workout",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "store.Workout": {
            "type": "object",
            "properties": {
                "calories_burned": {
                    "description": "in kcal",
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration_minutes": {
                    "type": "integer"
                },
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/store.WorkoutEntry"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "store.WorkoutEntry": {
            "type": "object",
            "properties": {
                "duration_seconds": {
                    "type": "integer"
                },
                "exercise_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "order_index": {
                    "type": "integer"
                },
                "reps": {
                    "type": "integer"
                },
                "sets": {
                    "type": "integer"
                },
                "weight": {
                    "description": "in kg",
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}